agribotbackend/requirements.txt
Django
django-cors-headers
djangorestframework
drf-yasg
pillow
django-decouple
dj-database-url
whitenoise
psycopg2-binary
gunicorn
django-unused-media
requests
google.generativeai
cloudinary
django-cloudinary-storage